Move `UnsupportedType` to parser crate
...in order to be reusable for type inference
